service-code-center-backend/pyproject.toml
[tool.poetry]
name = "data-scenario-center-backend"
version = "0.1.0"
description = ""
authors = ["ekcapaper <92367474+ekcapaper@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
aiofiles = "^24.1.0"
pyyaml = "^6.0.2"
watchfiles = "^0.24.0"
pydantic-settings = "^2.6.1"
loguru = "^0.7.2"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
